==> code/02_imputing_date_of_arrest.Rmd <==
---
title: "Imputing date of arrest"
author: "Martin Kosík"
date: "March 10, 2019"
output: github_document
editor_options: 
  chunk_output_type: console
---

```{r setup, echo=T, results='hide', message=F, warning=F}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(broom)
library(lubridate)
library(scales)
library(stargazer)
library(data.table)
library(here)
library(dqrng)
```

Set number of simulations
```{r}
number_of_simulations <- 5
```


Import and pre-process the data
```{r}
selected_vars <- c("person_id", "arest_date", "process_date")
memorial_lists <- fread(here::here("memo_list/memorial_lists.tsv"), encoding="UTF-8", sep="\t", 
                        select = selected_vars, quote="")

memorial_lists <- memorial_lists[, no_date := ifelse(process_date == "None" & arest_date == "None", 1, 0)]

memorial_lists <- memorial_lists %>% 
  separate(arest_date, sep = "\\.", into = c("DAY", "MONTH", "YEAR"), remove = FALSE) %>% 
  separate(process_date, sep = "\\.", into = c("DAY_PROCESS", "MONTH_PROCESS", "YEAR_PROCESS"), remove = FALSE) %>%
  mutate_at(c("DAY", "MONTH", "YEAR", "DAY_PROCESS", "MONTH_PROCESS", "YEAR_PROCESS"),
             funs(as.numeric(ifelse(. %in% c("None", "_"), NA, .))))

```

Get relative frequency of trials by months
```{r}
month_freq <- memorial_lists %>% 
  filter(!is.na(MONTH_PROCESS)) %>% 
  add_count() %>% 
  group_by(MONTH_PROCESS) %>% 
  summarise(month_freq = n()/n[1]) %>% 
  pull(month_freq)

```

We subset only the data where some date ,either of arrest or trial (called process in the data) is available.
We then sample month of trial where it is missing based in relative frequency of trials by month. 
```{r}
set.seed(2019)

some_date <-  memorial_lists %>% 
  filter(no_date == 0) %>% 
  mutate(MONTH_PROCESS_freq_imp = ifelse(is.na(MONTH_PROCESS),
                                           base::sample(1:12, size = 1, prob = month_freq),
                                           MONTH_PROCESS),
         DAY_PROCESS_freq_imp = ifelse(is.na(DAY_PROCESS),
                                           case_when(MONTH_PROCESS_freq_imp %in% c(1, 3, 5, 7, 8, 10, 12) ~ 
                                                       base::sample(1:31, size = 1, prob = rep(1/31, 31)),
                                                     MONTH_PROCESS_freq_imp %in% c(4, 6, 9, 11) ~ 
                                                       base::sample(1:30, size = 1, prob = rep(1/30, 30)),
                                                     MONTH_PROCESS_freq_imp == 2 ~
                                                       base::sample(1:28, size = 1, prob = rep(1/28, 28))),
                                           DAY_PROCESS))



```

Save the subset of the data where both date of arrest and date of trial are available
```{r}
both_dates <- some_date %>% 
  filter((!is.na(MONTH)) & (!is.na(MONTH_PROCESS))) %>% 
  mutate(date_arrest = parse_date_time(arest_date, "dmy"), 
         date_process = parse_date_time(process_date, "dmy")) %>% 
  filter(!(is.na(date_arrest) | is.na(date_process))) %>% 
  mutate(diff_days = interval(date_arrest, date_process) / ddays(1)) %>% 
  filter(YEAR_PROCESS >1920, YEAR_PROCESS < 1961) %>% 
  filter(diff_days >= 0)
```

Here we fit only simple log-linear model with only interept that we later plot over histogram.
```{r}
lm_model_simple <-  lm(log10(1 +diff_days) ~ 1, data = both_dates)
lm_model_simple_intercept <- lm_model_simple$coefficients[[1]]
lm_model_simple_sigma <- summary(lm_model_simple)$sigma
```

Below we fit the two-step model described in the paper
```{r}
glm.sign <- glm(I(diff_days > 0) ~ as.factor(YEAR_PROCESS), 
                data = both_dates, family = binomial(link = logit))

lm.ifpos <- lm(I(log(diff_days)) ~ as.factor(YEAR_PROCESS),
               data = both_dates, subset = diff_days > 0) 
```

Save the table with results of the model
```{r}
var_labels <- names(glm.sign$coefficients) %>% 
  str_replace_all("as.factor\\(YEAR_PROCESS\\)", "Year of Trial - ")


stargazer(glm.sign, lm.ifpos, 
          omit.stat = c("ser", "f"), label = "tab:date_imp_results", font.size = "small", no.space = T, 
          table.placement = "!h", single.row = T, header = F,
          covariate.labels = var_labels, title = "Arrest Date Imputation - Model Results",
          dep.var.labels = c("$I^y$", "$log(y^{\\text{pos}})$")) %>% 
 cat(sep = '\n', file = here::here("tables/date_imp_results.tex"))

```

Finally, we impute the missing values and save the results
```{r}
missing_date_of_arrest <- some_date %>% 
  filter(is.na(YEAR) & !is.na(YEAR_PROCESS)) %>% 
  dplyr::select(person_id, YEAR_PROCESS) %>% 
  filter(YEAR_PROCESS >1920, YEAR_PROCESS < 1961)

set.seed(2019)

missing_date_of_arrest <- missing_date_of_arrest %>% 
  mutate(diff_day_imp = round(exp(rnorm(n(), mean = lm_model_simple_intercept, sd = lm_model_simple_sigma)) - 1),
         diff_day_imp = ifelse(diff_day_imp < 0, 0, diff_day_imp) )


missing_date_of_arrest <- missing_date_of_arrest %>% 
  as_tibble() %>% 
  mutate(pred.sign = map(YEAR_PROCESS, ~ rbinom(number_of_simulations, 1, 
                                             predict(glm.sign, data.frame(YEAR_PROCESS = .), type="response")) ),   
         pred.pos.log =map(YEAR_PROCESS, ~ rnorm(number_of_simulations,
                                                 mean = predict(lm.ifpos, data.frame(YEAR_PROCESS = .)), 
                                                 sd = summary(lm.ifpos)$sigma)), 
         diff_day_imp_complex = round(exp(pred.pos.log) * pred.sign) )


pred.sign_prob <- predict(glm.sign, missing_date_of_arrest, type="response")
pred.sign <- matrix(NA, nrow = nrow(missing_date_of_arrest), ncol = number_of_simulations)

for(i in 1:nrow(pred.sign)){
  if(i %% 1000==0) {
      # Print on the screen some message
      cat(paste0("iteration: ", i, "\n"))
   }

  pred.sign[i, ] <- rbinom(number_of_simulations, 1, pred.sign_prob[i])
}



pred.pos.log_mean <- predict(lm.ifpos, missing_date_of_arrest, type="response")
pred.pos.log <- matrix(NA, nrow = nrow(missing_date_of_arrest), ncol = number_of_simulations)

for(i in 1:nrow(pred.pos.log)){
  if(i %% 1000==0) {
      # Print on the screen some message
      cat(paste0("iteration: ", i, "\n"))
   }
  pred.pos.log[i, ] <- rnorm(number_of_simulations, mean = pred.pos.log_mean[i], sd = summary(lm.ifpos)$sigma)
}

for(i in 1:nrow(pred.pos.log)){
  if(i %% 1000==0) {
      # Print on the screen some message
      cat(paste0("iteration: ", i, "\n"))
   }
  pred.pos.log[i, ] <- dqrnorm(number_of_simulations, mean = pred.pos.log_mean[i], sd = summary(lm.ifpos)$sigma)
}

saveRDS(pred.pos.log, here::here("data/pred_pos_log_matrix.RData"))
    
pred.pos.log <- readRDS(here::here("data/pred_pos_log_matrix.RData"))


diff_day_imp_complex <- round(exp(pred.pos.log) * pred.sign) 

diff_day_imp_complex <- as.data.frame(diff_day_imp_complex)

names(diff_day_imp_complex) <- paste0("diff_day_imp_complex_", 1:number_of_simulations)

missing_date_of_arrest <- cbind(missing_date_of_arrest, diff_day_imp_complex)

complex_imp <- some_date %>% 
  left_join(missing_date_of_arrest %>% dplyr::select(-YEAR_PROCESS), by = "person_id") %>% 
  melt(id.vars = c("person_id", "YEAR", "YEAR_PROCESS", "MONTH_PROCESS_freq_imp", "DAY_PROCESS_freq_imp"),
       measure.vars = paste0("diff_day_imp_complex_", 1:number_of_simulations),
       variable.name = "simulation_number", value.name = "diff_day_imp_complex")

rm(list = c("glm.sign", "lm.ifpos", "memorial_lists", "pred.pos.log", "pred.sign"))
gc()

complex_imp <- complex_imp %>% 
  mutate(date_process_freq_imp = as_date(str_c(YEAR_PROCESS, MONTH_PROCESS_freq_imp, DAY_PROCESS_freq_imp,
                                         sep = "-")),
         date_arrest_complex_imp = date_process_freq_imp - days(diff_day_imp_complex),
         year_arrest_complex_imp = year(date_arrest_complex_imp), 
         source_date_arrest = case_when(!is.na(YEAR) ~ "label",
                                        !is.na(year_arrest_complex_imp) ~ "prediction",
                                        TRUE ~ "missing"),
         YEAR = ifelse(!is.na(year_arrest_complex_imp), year_arrest_complex_imp, YEAR),
         simulation_number = str_c("YEAR_", str_sub(simulation_number, start = nchar("diff_day_imp_complex_") + 1))) 


complex_imp <- complex_imp %>% 
  dcast(person_id + source_date_arrest ~ simulation_number, value.var = "YEAR")



imputation <- complex_imp %>% 
  filter(source_date_arrest != "missing") %>% 
  dplyr::select(person_id, starts_with("YEAR_"), source_date_arrest)

saveRDS(imputation, file = here::here("data/arrest_date_imputation.RData"))


```

## Plots 
```{r results="hide", echo = TRUE, fig.show='hide'}
mylog10_trans <- function (base = 10) 
{
  trans <- function(x) log(x + 1, base)
  inv <- function(x) base^(x) - 1 
  trans_new(paste0("log-", format(base)), trans, inv, log_breaks(base = base), 
            domain = c(0, Inf))
}


missing_date_of_arrest %>% 
  mutate(log_day = log(1 + diff_day_imp_complex)) %>% 
  ggplot(aes(x = diff_day_imp_complex)) + geom_histogram(aes(y=..density..),col = "white") + theme_minimal()+
  scale_x_continuous(trans = "mylog10", breaks = c(0,2,10,100,1000,10000))+
  labs(x = "Number of days between arrest and trail") + 
  theme(axis.line = element_line(size = 1), 
       # panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank(), 
        text = element_text(size=12)) + expand_limits(x = -0.01)
ggsave(here::here("plots/imputing_arrest_date/mixed_model_preds_hist.pdf"))

 
both_dates %>% 
  #filter(abs(diff_days) < 5000) %>% 
  mutate(log_diff_days = log(1 + diff_days)) %>% 
  ggplot(aes(x = diff_days)) + geom_histogram(aes(y=..density..), col = "white") + 
    #geom_density(aes(y=..density.., position="stack"), col = "red") +
    stat_function(fun = function(x, mean, sd) dnorm(log10(x + 1) , mean, sd), 
                  args = list(mean = lm_model_simple$coefficients[1], sd = summary(lm_model_simple)$sigma), 
                  col = "blue", size=1.5) + 
  scale_x_continuous(trans = "mylog10", breaks = c(0,2,10,100,1000,10000)) +
  theme_minimal() + labs(x = "Number of days between arrest and trial") +
  theme(axis.line = element_line(size = 1), 
       # panel.grid.major.x = element_blank(), 
        panel.grid.minor.x = element_blank(), 
        text = element_text(size=12))

ggsave(here::here("plots/imputing_arrest_date/simple_lm_hist.pdf"))
```
