==> code/03_border_frontiers.Rmd <==
---
title: "Border Frontiers"
author: "Martin Kosík"
date: "May 1, 2019"
output: github_document
editor_options: 
  chunk_output_type: console
---

```{r setup, echo=T, results='hide', message=F, warning=F}
knitr::opts_chunk$set(echo = TRUE, fig.show = 'hide')
library(data.table)
library(tidyverse)
library(here)
library(sf)
Sys.setlocale("LC_CTYPE", "russian") 
library(mapview)
# library(mapedit) (if you want to draw polygons yourself)


library(stringi)
```

Load the geocoded dataset from Zhukov and Talibova (2018) and our Memorial data. 
```{r}
selected_vars <- c("person_id", "nation", "arest_date", "process_date","live_place", 
                   "birth_place", "birth_year",  "surname", "name", "patronimic", "memory_book")

memorial_lists <- fread(here::here("memo_list/memorial_lists.tsv"), encoding="UTF-8", sep="\t", 
                        select = selected_vars, quote="", na.strings = "None")


memorial_lists <- memorial_lists %>% 
  mutate(live_or_birth_place = ifelse(is.na(live_place), birth_place, live_place),
         live_or_birth_place_lat = stri_trans_general(live_or_birth_place, "russian-latin/bgn")
         )

memorial_lists %>% 
  count(birth_place, sort = T)

head(events$ORIGIN_NAME)

load(here::here("data/eventsClean_v1.RData"))

zhukov_locations <- events %>% 
  count(ORIGIN_NAME, ORIGIN_LONG, ORIGIN_LAT) %>% 
  mutate(ORIGIN_NAME_lat = stri_trans_general(ORIGIN_NAME, "russian-latin/bgn"))

rm(events)
gc()

head(memorial_lists)

memorial_locations <- memorial_lists %>%
  count(live_or_birth_place_lat)

matched_locations <- memorial_locations %>% 
  inner_join(zhukov_location, by = c("live_or_birth_place_lat" = "ORIGIN_NAME_lat"))


events <- events %>% 
  dplyr::select(YEAR_STATUTE, YEAR, DATE, ORIGIN_LONG, ORIGIN_LAT, NAME, BIRTH_DATE, SOURCE)

events <- events %>% 
  mutate(names_split = str_split(NAME, pattern = " "),
         first_name = as.factor(map_chr(names_split, nth, 2)),
         surname = as.factor(map_chr(names_split, 1)),
         patronymic = as.factor(map_chr(names_split, nth, 3)),
         number_of_names = as.factor(map_int(names_split, length))) %>% 
  dplyr::select(-c(names_split, NAME))

saveRDS(events, file = here::here("data/eventsClean_v2.RData"))

events <- readRDS(here::here("data/eventsClean_v2.RData"))

events <- events %>% 
  mutate(BIRTH_DATE = as.numeric(BIRTH_DATE))
```

We drop possible duplicates. 
```{r}
events_distinct <- events %>% 
  distinct(surname, first_name, BIRTH_DATE, patronymic, SOURCE, ORIGIN_LONG, ORIGIN_LAT)

```

We merge Zhukov and Talibova (2018) dataset to our data based names, birth year, source.  
```{r}
merged <- memorial_lists %>% 
  left_join(events_distinct, by = c("surname", "name" = "first_name", "birth_year" = "BIRTH_DATE",
                           "patronimic" = "patronymic", "memory_book" = "SOURCE"))
```

We  drop observations with no or multiple matches. 
```{r}
merged <- merged %>% 
  dplyr::select(person_id, ORIGIN_LONG, ORIGIN_LAT) %>% 
  filter(!is.na(ORIGIN_LONG), !is.na(ORIGIN_LAT))

rm(list = c("events", "events_distinct", "memorial_lists"))
gc()

multiple_matches <- merged %>% 
  count(person_id, sort = T) %>% 
  filter(n > 1)

merged <- merged %>% 
  anti_join(multiple_matches, by = "person_id")

rm(multiple_matches)
```

We load the shape of USSR borders and convert it to Multiline object (instead of multiple polygon object). 
```{r}
ussr_map <- st_read(here::here("data/ussr_shapefiles/1926SovietUnion.shp"))

st_crs(ussr_map) <- "+proj=longlat +datum=WGS84 +no_defs"

ussr_borders <-  ussr_map %>% 
    summarise(geometry = st_union(.)) %>% 
  st_boundary()

ussr_map <- ussr_map %>% 
  dplyr::select(Id, NameENG, NameRUS)
```

In the following code chunk, I drew polygons along the Soviet border with other countries which I then interest with the Multiline object of Soviet boundaries. If I would not do this, the coastlines would be treated as borders too. Thus, do not run this code chunk (unless you want draw the polygons yourself).
```{r eval=FALSE}

editMap(mapview(ussr_map))

drawed_polygons <- editMap(mapview(ussr_borders))


saveRDS(drawed_polygons, file = here::here("data/drawed_polygons.RData"))
```


```{r eval=FALSE}
drawed_polygons_only_western_border <- editMap(mapview(ussr_borders))


saveRDS(drawed_polygons_only_western_border, file = here::here("data/drawed_polygons_only_western_border.RData"))

```


```{r}
drawed_polygons <- readRDS(here::here("data/drawed_polygons.RData"))

# you can check the drawed polygons with by runnung mapview(drawed_polygons$drawn)

border_lines <- st_intersection(ussr_borders, drawed_polygons$drawn)


buffer_250km <- border_lines %>% 
  st_transform(crs = 7801) %>% 
  st_buffer(dist = 250000, nQuadSegs = 1) %>% 
  st_intersection(ussr_map %>% st_transform(crs = 7801)) %>% 
  st_transform(crs = "+proj=longlat +datum=WGS84 +no_defs")



ggplot(ussr_map) +
    geom_sf(fill = "gray95", color = "gray50", size = 0.5) +
    geom_sf(data = st_transform(buffer_250km, crs = "+proj=longlat +datum=WGS84 +no_defs"), fill = "gray50")+
    theme_void() +
    coord_sf(crs = "+proj=longlat +datum=WGS84 +no_defs", ndiscr = F) + 
    labs(fill = "Border province (in Europe)") + 
    theme(legend.position="bottom")

ggsave(here::here("plots/final/border_buffer_map.pdf"))


```


```{r}
arrests_sf <- st_as_sf(merged, coords = c("ORIGIN_LONG", "ORIGIN_LAT"), crs = 4326) 

joined <-  arrests_sf %>%
  st_join(buffer_250km, suffix = c("", "_buffer")) %>% 
   st_join(ussr_map, suffix = c("", "_whole")) 


joined_df <- joined %>% 
  st_set_geometry(NULL) %>% 
  mutate(location = case_when(!is.na(NameENG) ~ "within 250 km border buffer", 
                              !is.na(NameENG_whole) ~ "outside 250 km border buffer",
                              TRUE ~ "outside USSR"))


joined_df <- joined_df %>% 
  dplyr::select(person_id, location)

fwrite(joined_df, here::here("data/border_arrests.csv"))

```

```{r}
library(scales)
ussr_map <- st_read(here::here("data/ussr_shapefiles/1926SovietUnion.shp")) %>% 
  mutate(share_german = Germans/PopALL,
         germans_cat = cut(Germans, breaks = c(0, 1000, 10000, 100000, 400000)),
         share_german_cat = cut(share_german, breaks = c(0, 0.001, 0.01, 0.1, 0.7)))


st_crs(ussr_map) <- "+proj=longlat +datum=WGS84 +no_defs"


ggplot(ussr_map) +
    geom_sf(aes(fill = Germans), color = "gray50", size = 0.5) +
    scale_fill_continuous(low = "white", high = "red", 
                          labels = function(x) format(x, big.mark = ",", scientific = FALSE)) +
    guides(fill = guide_colourbar(barwidth = 20, barheight = 1.5)) + 
    theme_void() +
    coord_sf(crs = "+proj=longlat +datum=WGS84 +no_defs", ndiscr = F) + 
    labs(fill = "Number of Germans"
         #,caption = "The source of the data is the 1926 Soviet Census"
         ) +
    theme(legend.position="bottom")

ggplot(ussr_map) +
    geom_sf(aes(fill = germans_cat), color = "gray50", size = 0.5) +
    scale_fill_brewer(palette = "Reds", 
                          labels = c("0 to 1000", "1001 to 10,000", "10,001 to 100,000", "100,001 to 400,000")) +
    #guides(fill = guide_colourbar(barwidth = 20, barheight = 1.5)) + 
    theme_void() +
    coord_sf(crs = "+proj=longlat +datum=WGS84 +no_defs", ndiscr = F) + 
    labs(fill = "Number of Germans"
         #,caption = "The source of the data is the 1926 Soviet Census"
         ) +
    theme(legend.position="bottom")


ggsave(here::here("plots/for_presentation/map_number_of_germans_discrete.pdf"), scale = 0.75)

ggplot(ussr_map) +
    geom_sf(aes(fill = share_german), color = "gray50", size = 0.5) +
    scale_fill_continuous(low = "white", high = "red", 
                          labels = percent_format(accuracy = 0.001),
                          #limits = c(0.000000001, 1),
                          trans = "log")+
    guides(fill = guide_colourbar(barwidth = 20, barheight = 1.5)) + 
    theme_void() +
    coord_sf(crs = "+proj=longlat +datum=WGS84 +no_defs", ndiscr = F) + 
    labs(fill = "Share of Germans"
         #,caption = "The source of the data is the 1926 Soviet Census"
         ) +
    theme(legend.position="bottom")
ggsave(here::here("plots/for_presentation/map_share_of_germans_log_scale.pdf"), scale = 0.7)

ggplot(ussr_map) +
    geom_sf(aes(fill = share_german), color = "gray50", size = 0.5) +
    scale_fill_continuous(low = "white", high = "red", 
                          labels = percent_format(accuracy = 0.001)) +
                          #limits = c(0.000000001, 1))+
    guides(fill = guide_colourbar(barwidth = 20, barheight = 1.5)) + 
    theme_void() +
    coord_sf(crs = "+proj=longlat +datum=WGS84 +no_defs", ndiscr = F) + 
    labs(fill = "Share of Germans"
         #,caption = "The source of the data is the 1926 Soviet Census"
         ) +
    theme(legend.position="bottom")
ggsave(here::here("plots/for_presentation/map_share_of_germans_log_scale.pdf"), scale = 0.7)


ggplot(ussr_map) +
    geom_sf(aes(fill = share_german_cat), color = "gray50", size = 0.5) +
    scale_fill_brewer(palette = "Reds",
                         labels = c("0 to 0.1", "0.1 to 1", "1 to 10", "10 to 70")
                          ) +
    #guides(fill = guide_colourbar(barwidth = 20, barheight = 1.5)) + 
    theme_void() +
    coord_sf(crs = "+proj=longlat +datum=WGS84 +no_defs", ndiscr = F) + 
    labs(fill = "Share of Germans \n (in percent)"
         #,caption = "The source of the data is the 1926 Soviet Census"
         ) +
    theme(legend.position="bottom")


ggsave(here::here("plots/for_presentation/map_share_of_germans_discrete.pdf"), scale = 0.75)

```

### Only western border
```{r}
drawed_polygons <- readRDS(here::here("data/drawed_polygons_only_western_border.RData"))

# you can check the drawed polygons by runnung mapview(drawed_polygons$drawn)

border_lines <- st_intersection(ussr_borders, drawed_polygons$drawn)


buffer_250km <- border_lines %>% 
  st_transform(crs = 7801) %>% 
  st_buffer(dist = 250000, nQuadSegs = 1) %>% 
  st_intersection(ussr_map %>% st_transform(crs = 7801)) %>% 
  st_transform(crs = "+proj=longlat +datum=WGS84 +no_defs")



ggplot(ussr_map) +
    geom_sf(fill = "gray95", color = "gray50", size = 0.5) +
    geom_sf(data = st_transform(buffer_250km, crs = "+proj=longlat +datum=WGS84 +no_defs"), fill = "gray50")+
    theme_void() +
    coord_sf(crs = "+proj=longlat +datum=WGS84 +no_defs", ndiscr = F) + 
    labs(fill = "Border province (in Europe)") + 
    theme(legend.position="bottom")

ggsave(here::here("plots/for_presentation/border_buffer_map_only_western_border.pdf"))

```

```{r}
arrests_sf <- st_as_sf(merged, coords = c("ORIGIN_LONG", "ORIGIN_LAT"), crs = 4326) 

joined <-  arrests_sf %>%
  st_join(buffer_250km, suffix = c("", "_buffer")) %>% 
   st_join(ussr_map, suffix = c("", "_whole")) 


joined_df <- joined %>% 
  st_set_geometry(NULL) %>% 
  mutate(location = case_when(!is.na(NameENG) ~ "within 250 km of western border buffer", 
                              !is.na(NameENG_whole) ~ "outside 250 km of western border buffer",
                              TRUE ~ "outside USSR"))


joined_df <- joined_df %>% 
  dplyr::select(person_id, location)

fwrite(joined_df, here::here("data/border_arrests_western.csv"))

```
