==> code/07_diff_in_diff_reg_tables.Rmd <==
---
title: "Regression tables"
author: "Martin Kosík"
date: "21 dubna 2019"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(stargazer)
```

```{r}
geopol_vars <- str_subset(names(min_by_year), "geopol")

fmla_pred_full_imp_date_no_trends_geopol <- as.formula(paste("log_n_pred_full_imp_date ~ ", 
                         paste0("german:","year_", years, collapse = " + "),  "+ as.factor(YEAR) +
                         ethnicity+ ",  paste0(geopol_vars, collapse = " + ")))

fmla_pred_full_imp_date_no_trends_geopol_rehab <- as.formula(paste("log_n_pred_full_imp_date_rehab ~ ", 
                         paste0("german:","year_", years, collapse = " + "),  "+ as.factor(YEAR) +
                         ethnicity+ ",  paste0(geopol_vars, collapse = " + ")))

fmla_pred_full_imp_date_no_trends <- as.formula(paste("log_n_pred_full_imp_date ~ ", 
                         paste0("german:","year_", years, collapse = " + "),  "+ as.factor(YEAR) +
                         ethnicity"))
```


```{r}
default_model <-  lm(fmla_pred_full_imp_date_no_trends_geopol, data = min_by_year)
rehabs_model <-  lm(fmla_pred_full_imp_date_no_trends_geopol_rehab, data = min_by_year)
no_ind_ethn_data <- min_by_year %>% 
  filter(ind_country == 0 | ethnicity == "German") 

no_ind_ethn_model <-  lm(data = no_ind_ethn_data, formula = fmla_pred_full_imp_date_no_trends)


robust_se <-  function(model,  vcov = "CR2",  level = 0.95, data = min_by_year){
  model %>% 
  conf_int(vcov = vcov,  level = level,
          cluster = data$ethnicity,  test = "Satterthwaite") %>% 
    pull(SE)
}

default_robust_se <- robust_se(default_model)
rehabs_robust_se <- robust_se(rehabs_model)
no_ind_ethn_robust_se <- robust_se(no_ind_ethn_model, data = no_ind_ethn_data)


```


```{r}
var_labels <- str_c("$",names(default_model$coefficients)) %>% 
  str_subset("german:year_") %>% 
  #str_replace_all("t", "t\\\\") %>% 
  str_replace_all("\\_", "-") %>% 
  str_replace(":", "*") %>% 
  str_c("$")

var_labels <- str_c("$",names(default_model$coefficients)) %>% 
  str_subset("german:year_") %>% 
  str_replace("german:year_", "\\\\text{German}\\_{i} \\\\cdot \\\text{Year}\\_{t}^") %>% 
  str_c("$")


var_labels <- names(default_model$coefficients) %>% 
  str_subset("german:year_") %>% 
  #str_replace_all("t", "t\\\\") %>% 
  str_replace_all("\\_", "-") %>% 
  str_replace(":", "*")




var_labels <- str_c("$",names(default_model$coefficients)) %>% 
  str_subset("german:year_") %>% 
  str_replace("german:year_", "\\\\beta\\_{") %>% 
  str_c("}$")



diff_table <-  stargazer(default_model, no_ind_ethn_model, rehabs_model, 
          keep = "german:year_+", omit.stat = c("ser", "f", "rsq"), label = "dif_table",
          se = list(default_robust_se, no_ind_ethn_robust_se, rehabs_robust_se), 
          covariate.labels = var_labels, title = "Difference-in-differences results",
          dep.var.labels.include = FALSE, single.row = T, header = F,
          font.size = "small", no.space = T,  dep.var.caption  = "Model",
          dep.var.labels = c("$\\log(1 + y_{it})$", "$\\log(1 + y_{it})$"), 
          add.lines = list(#c("Year dummies", "Yes", "Yes", "Yes"),
                           #c("Ethnicity dummies", "Yes", "Yes", "Yes"),
                           c("Eth. with ind. state excluded", "No", "Yes", "No"),
                           c("Only rehabilitated ind.", "No", "No", "Yes"), 
                           c("Geopol. relations controls", "Yes", "No", "Yes"),
                           c("Ethnicity-spec. time trends", "No", "No", "No")),
          #column.labels = c("Default", "Ind. eth. excluded", "Only Rehabs"),
          #star.cutoffs = c(0.05, 0.01, 0.001),
          table.placement = "!h",
          notes = "This will be replaced", notes.append = FALSE, notes.align = "l")

note_latex <- "\\multicolumn{4}{l} {\\parbox[t]{\\textwidth}{\\textit{Notes:} Cluster-robust standard errors are in the parentheses. The coefficients from model (1) are plotted in the figure \\ref{fig:did_effets}, from model (2) in the figure \\ref{fig:did_effets_no_ind_countries}, and from model (3) in the figure \\ref{fig:did_effects_rehabs}. For additional information, refer to the notes of the respective figures. $^{*}$p$<$0.1; $^{**}$p$<$0.05; $^{***}$p$<$0.01}}"

diff_table[grepl("Note",diff_table)] <- note_latex

diff_table %>% 
  cat(sep = '\n', file = here::here("tables/diff_table.tex"))
```


## Time trends 
```{r}
geopol_vars <- str_subset(names(min_by_year), "geopol")

fmla_pred_full_imp_date_lin_trends_geopol <- as.formula(paste("log_n_pred_full_imp_date ~ ", 
                         paste0("german:","year_", years, collapse = " + "),  "+ as.factor(YEAR) +
                         ethnicity + ethnicity: YEAR +", paste0(geopol_vars, collapse = " + ")))

fmla_pred_full_imp_date_geopol <- as.formula(paste("log_n_pred_full_imp_date ~ ", 
                         paste0("german:","year_", years, collapse = " + "),  "+ as.factor(YEAR) +
                         ethnicity + ethnicity: YEAR+ ethnicity: YEAR_sq + ",  paste0(geopol_vars, collapse = " + ")))


fmla_pred_full_imp_date_no_trends_geopol <- as.formula(paste("log_n_pred_full_imp_date ~ ", 
                         paste0("german:","year_", years, collapse = " + "),  "+ as.factor(YEAR) +
                         ethnicity+ ",  paste0(geopol_vars, collapse = " + ")))
```


```{r}
default_model <-  lm(fmla_pred_full_imp_date_no_trends_geopol, data = min_by_year)
lin_trends_model <-  lm(fmla_pred_full_imp_date_lin_trends_geopol, data = min_by_year)
quad_trends_model <-  lm(fmla_pred_full_imp_date_lin_trends_geopol, data = min_by_year)

robust_se_list <- map(list(default_model, lin_trends_model, quad_trends_model), robust_se)

```

```{r}
diff_table2 <-  stargazer(default_model, lin_trends_model, quad_trends_model, 
          keep = "german:year_+", omit.stat = c("ser", "f", "rsq"), label = "dif_table_trends",
          se = robust_se_list, 
          covariate.labels = var_labels, title = "Difference-in-differences results - Ethnicity-specific time trends",
          dep.var.labels.include = FALSE, single.row = T, header = F,
          font.size = "small", no.space = T, 
          dep.var.labels = c("$\\log(1 + y_{it})$"),  dep.var.caption  = "Model",
          add.lines = list(#c("Year dummies", "Yes", "Yes", "Yes"),
                           #c("Ethnicity dummies", "Yes", "Yes", "Yes"),
                           c("Ethnicity-spec. time trends", "None", "Linear", "Quadratic"),
                           c("Eth. with ind. state excluded", "No", "Yes", "No"),
                         #  c("Only rehabil. ind.", "No", "No", "Yes"), 
                           c("Geopol. relations controls", "Yes", "Yes", "Yes")),
          # column.labels = c("Default", "Ind. eth. excluded", "Only Rehabs"),
         # star.cutoffs = c(0.05, 0.01, 0.001),
          table.placement = "!h", 
         notes = "This will be replaced", notes.append = FALSE, notes.align = "l")

note_latex2 <- "\\multicolumn{4}{l} {\\parbox[t]{\\textwidth}{\\textit{Notes:} Cluster-robust standard errors are in the parentheses. The coefficients from these models are plotted in the figure \\ref{fig:did_robustness_time_trends}. For additional information, refer to the notes of the figures \\ref{fig:did_robustness_time_trends}. $^{*}$p$<$0.1; $^{**}$p$<$0.05; $^{***}$p$<$0.01}}"

diff_table2[grepl("Note",diff_table2)] <- note_latex2

diff_table2 %>% 
  cat(sep = '\n', file = here::here("tables/diff_table_trends.tex"))
```

## Ethnicity imputation adjustments

```{r}
geopol_vars <- str_subset(names(min_by_year), "geopol")

fmla_pred_imp_date_no_trends_geopol <- as.formula(paste("log_n_pred_imp_date ~ ", 
                         paste0("german:","year_", years, collapse = " + "),  "+ as.factor(YEAR) +
                         ethnicity+ ",  paste0(geopol_vars, collapse = " + ")))

fmla_pred_full_imp_date_no_trends_geopol <- as.formula(paste("log_n_pred_full_imp_date ~ ", 
                         paste0("german:","year_", years, collapse = " + "),  "+ as.factor(YEAR) +
                         ethnicity+ ",  paste0(geopol_vars, collapse = " + ")))

fmla_pars_pred_imp_date_no_trends_geopol <- as.formula(paste("log_n_pred_pars_imp_date ~ ", 
                         paste0("german:","year_", years, collapse = " + "),  "+ as.factor(YEAR) +
                         ethnicity+ ",  paste0(geopol_vars, collapse = " + ")))
```

```{r}

pred_adj_formulas_full_years_geopol <- c(fmla_pred_full_imp_date_no_trends_geopol,
                                       fmla_pars_pred_imp_date_no_trends_geopol,
                                       fmla_pred_imp_date_no_trends_geopol)


pred_adj_formulas_full_years_geopol_lm<-map(pred_adj_formulas_full_years_geopol, ~lm(formula = ., data = min_by_year))

robust_se_list_pred_adj <- map(pred_adj_formulas_full_years_geopol_lm, robust_se)
```

```{r}
diff_table3 <-  stargazer(pred_adj_formulas_full_years_geopol_lm, 
          keep = "german:year_+", omit.stat = c("ser", "f", "rsq"), label = "dif_table_pred_adj",
          se = robust_se_list_pred_adj, 
          covariate.labels = var_labels, title = "Difference-in-differences results - Ethnicity Imputation Adjustments",
          dep.var.labels.include = FALSE, single.row = T, header = F,
          font.size = "small", no.space = T, dep.var.caption  = "Model",
         # dep.var.labels = c("$\\log(1 + y_{it})$"), 
          add.lines = list(#c("Year dummies", "Yes", "Yes", "Yes"),
                           c("Ethnicity imputation adjust.", "Full-matrix", "Parsimonoius", "None"),
                           c("Ethnicity-spec. time trends", "None", "None", "None"),
                           c("Eth. with ind. state excluded", "No", "No", "No"),
                         #  c("Only rehabil. ind.", "No", "No", "Yes"), 
                           c("Geopol. relations controls", "Yes", "Yes", "Yes")),
          # column.labels = c("Default", "Ind. eth. excluded", "Only Rehabs"),
         # star.cutoffs = c(0.05, 0.01, 0.001),
          table.placement = "!h", 
         notes = "This will be replaced", notes.append = FALSE, notes.align = "l")

note_latex3 <- "\\multicolumn{4}{l} {\\parbox[t]{\\textwidth}{\\textit{Notes:} Cluster-robust standard errors are in the parentheses. The coefficients from these models are plotted in the figure \\ref{fig:did_robustness_pred_adj}. For additional information, refer to the notes of the figures \\ref{fig:did_robustness_pred_adj}. $^{*}$p$<$0.1; $^{**}$p$<$0.05; $^{***}$p$<$0.01}}"

diff_table3[grepl("Note",diff_table3)] <- note_latex3

diff_table3 %>% 
  cat(sep = '\n', file = here::here("tables/diff_table_pred_adj.tex"))

```



```{r}
stargazer(diff_model, diff_model_restr,diff_model_no_log,
          se = starprep(diff_model_robust, diff_model_restr_robust, diff_model_no_log_robust),
          p = starprep(diff_model_robust, diff_model_restr_robust, diff_model_no_log_robust, stat = "p.value"),
          keep = "german:post_+", omit.stat = c("ser", "f"), label = "dif_table",
          covariate.labels = var_labels, title = "Difference-in-differences results",
          dep.var.labels = c("$\\log(1 + y_{it})$", "$y_{it}$"), 
          add.lines = list(c("Year dummies", "Yes", "Yes", "Yes"),
                           c("Ethnicity dummies", "Yes", "Yes", "Yes")),
          column.labels = c("1921-1958", "1921-1945", "1921-1958")) %>% 
 cat(sep = '\n', file = here("tables/diff_table.tex"))



stargazer(glm.sign, lm.ifpos, 
          omit.stat = c("ser", "f"), label = "tab:date_imp_results", font.size = "small", no.space = T, 
          table.placement = "!h", single.row = T, header = F,
          covariate.labels = var_labels, title = "Arrest Date Imputation - Model Results",
          dep.var.labels = c("$I^y$", "$log(y^{\\text{pos}})$")) %>% 
 cat(sep = '\n', file = here::here("tables/date_imp_results.tex"))

```
