==> code/09_synthetic_control_method.Rmd <==
---
title: "Synthetic control"
author: "Martin Kosík"
date: "February 14, 2019"
output: github_document
editor_options: 
  chunk_output_type: console
---

```{r setup, echo=T, results='hide', message=F, warning=F}
knitr::opts_chunk$set(echo = TRUE,  fig.show = 'hide')
library(tidyverse)
library(broom)
library(lubridate)
library(here)
library(readxl)
library(MSCMT)
library(kableExtra)
library(knitr)
library(gghighlight)
library(timetk)
library(scales)
source(here::here("code/functions.R"))
```

```{r}
ethnicity_controls <- read_excel(here::here("data/ethnicity_info.xlsx")) %>%
  mutate(ethnicity_id = as.numeric(1:n()),
         urb_rate_pct = urb_rate * 100)
  
min_by_year <-  read_csv(here::here("data/min_by_year_preds.csv")) %>% 
  mutate(log_n = log(1 + label ),
         log_n_pred_full = log(1 + label + pred_adj_full_scaled),
         log_n_pred = log(1 + label + prediction),
         log_n_imp_date = log(1 + label_imp_date),
         log_n_pred_full_imp_date = log(1 + label_imp_date + pred_adj_full_scaled_imp_date),
         n_pred_full_imp_date = label_imp_date + pred_adj_full_scaled_imp_date,
         log_n_pred_full_imp_date_rehab = log(1 + label_rehab + pred_adj_full_scaled_rehab)) %>% 
  left_join(ethnicity_controls, by = "ethnicity") %>% 
  mutate(german = ifelse(ethnicity == "German", 1, 0),
         post_german = german * ifelse(YEAR >= 1933, 1, 0)) 

```

```{r}
opts_current$set(label = "sc_predictors")
ethnicity_controls %>% 
  arrange(ethnicity) %>% 
  dplyr::select(ethnicity, pop_total, clad_sim, urb_rate_pct, ind_country) %>% 
  kable("latex", booktabs = T, digits = 2, linesep = "", format.args = list(big.mark = " "),
        col.names = c("Ethnic group", "Total population", "Ling. similarity to Russian", 
                      "Urbanization rate", "Ind. state"),
        caption = "Pre-treatment characteristics of ethnic groups in the USSR") %>%
  footnote(general = "Total population and urbanization rate of the ethnic group in the USSR is taken from 1926 census. The linguistic similarity to Russian is measured by the number of common nodes in the language tree (cladistic similarity). Independent state equals one if the ethnic group was a core group in an independent country that existed in the interwar period.",
           threeparttable = T) %>% 
  kable_styling(latex_options = c("hold_position")) %>% 
    write_file(here::here("tables/sc_predictors.tex"))


```

Data preparation
```{r data preparation}
data_prep_mscmt <- listFromLong(as.data.frame(min_by_year), unit.variable = "ethnicity_id", 
                                time.variable="YEAR", unit.names.variable="ethnicity")

```

Definition of the variables
```{r definiton of variables}
dep_var <- "log_n_pred_full_imp_date"

treatment.identifier <- "German"
controls.identifier  <- setdiff(colnames(data_prep_mscmt[[1]]),
                                 treatment.identifier)
times.dep  <- cbind("log_n_pred_full_imp_date" = c("1921","1932"))


times.pred <- cbind("log_n_pred_full_imp_date" = c("1921","1932"),
                    "pop_total"                = c("1921","1932"),
                    "clad_sim"                 = c("1921","1932"),
                    "urb_rate"                 = c("1921","1932"))

agg.fns <- rep("id", ncol(times.pred))
```


```{r}
sc_results <- mscmt(data_prep_mscmt, treatment.identifier, controls.identifier, times.dep, times.pred, agg.fns, seed=2019)


dep_var <- "log_n_pred_full_imp_date"
times.dep  <- cbind("log_n_pred_full_imp_date" = c("1921","1932"))

sc_placebo_pred_full_imp_date <- mscmt(data_prep_mscmt, treatment.identifier, controls.identifier, times.dep,
                                       times.pred, agg.fns, seed=2019, placebo = TRUE, single.v=TRUE)


agg.fns <- rep("mean", ncol(times.pred))                       
dep_var <- "log_n_pred_full_imp_date"
times.dep  <- cbind("log_n_pred_full_imp_date" = c("1921","1932"))

sc_placebo_pred_full_imp_date_mean <- mscmt(data_prep_mscmt, treatment.identifier, controls.identifier, times.dep, times.pred, agg.fns, seed=2019, placebo = TRUE, single.v=TRUE)

sc_results_mean <- mscmt(data_prep_mscmt, treatment.identifier, controls.identifier, times.dep, times.pred, agg.fns, seed=2019)

```

```{r tables of predictor means}
mean_all <-  min_by_year %>% 
  filter(YEAR < 1933) %>% 
  summarize_at(c("log_n_pred_full_imp_date", "pop_total", "urb_rate_pct", "clad_sim"), mean) %>% 
  gather(key = "variable", value = "all_groups")

ethnicity_controls %>% 
  summarize_all(mean)


opts_current$set(label = "sc_predictor_means")
sc_results$combined %>% 
  map(tk_tbl) %>% 
  bind_rows(.id = "variable") %>% 
  filter(variable %in% c(dep_var,"pop_total", "clad_sim", "urb_rate_pct")) %>% 
  filter(index < 1933) %>% 
  group_by(variable) %>% 
  summarize_all(mean) %>% 
  select(-c(index, gaps)) %>% 
  left_join(mean_all, by = "variable") %>% 
  mutate(variable = recode_factor(variable, "log_n_pred_full_imp_date" =
                                    "Log(1 + arrests)", 
                                  "pop_total" = "Total population",
                                  "clad_sim" = "Ling. similarity to Russian",
                                  "urb_rate_pct" = "Urbanization rate", .ordered = TRUE) )%>% 
  slice(c(2, 3, 4, 1)) %>% 
  kable("latex", booktabs = T, digits = 2, linesep = "", format.args = list(big.mark = " "),
        col.names = c("Variable", "Actual", "Synthetic", 
                      "Mean of all 38 ethnicities"),
        caption = "Pre-treatment Predictor Means") %>%
  add_header_above(c(" " = 1, "German minority" = 2)) %>% 
  footnote(general = "Log(1 + arrests) is averaged over the pre-treatment period (1921-1932). All other predictor are time-invariant. Total population and urbanization rate are taken from 1926 Soviet census.",
           threeparttable = T) %>% 
  kable_styling(latex_options = c("hold_position")) %>% 
  write_file(here::here("tables/sc_predictor_means.tex"))  

v_weights <-  as_tibble(sc_results_mean$v, rownames = "variable") %>% 
  mutate(variable = str_sub(variable, end = -16),
         variable =  ifelse(variable == "urb_rate", "urb_rate_pct", variable)) %>% 
  dplyr::select(variable, v_weight = max.order)


mean_all <- mean_all %>% 
  left_join(v_weights, by = "variable")
  

  
opts_current$set(label = "sc_predictor_means_robustness")
sc_results_mean$combined %>% 
  map(tk_tbl) %>% 
  bind_rows(.id = "variable") %>% 
  filter(variable %in% c(dep_var,"pop_total", "clad_sim", "urb_rate_pct")) %>% 
  filter(index < 1933) %>% 
  group_by(variable) %>% 
  summarize_all(mean) %>% 
  select(-c(index, gaps)) %>% 
  left_join(mean_all, by = "variable") %>% 
  mutate(variable = recode_factor(variable, "log_n_pred_full_imp_date" =
                                    "Log(1 + arrests)", 
                                  "pop_total" = "Total population",
                                  "clad_sim" = "Ling. similarity to Russian",
                                  "urb_rate_pct" = "Urbanization rate", .ordered = TRUE) )%>% 
  slice(c(2, 3, 4, 1)) %>% 
  kable("latex", booktabs = T, digits = c(2, 2, 2, 3), linesep = "", format.args = list(big.mark = " "),
        col.names = c("Variable", "Actual", "Synthetic", 
                      "Mean of all ethnicities", "$V$ weights"), 
        caption = "Pre-treatment Predictor Means", escape = TRUE) %>%
  add_header_above(c(" " = 1, "German minority" = 2)) %>% 
  footnote(general = "Log(1 + arrests) is averaged over the pre-treatment period (1921-1932). All other predictor are time-invariant. Total population and urbanization rate are taken from 1926 Soviet census.",
           threeparttable = T) %>% 
  kable_styling(latex_options = c("hold_position"), font_size = 10) %>% 
  write_file(here::here("tables/sc_predictor_means_robustness.tex"))  

```


```{r placebo highlight function}
placebo_highlight_all(sc_placebo_pred_full_imp_date, "log_n_pred_full_imp_date")
ggsave(here::here("plots/synthetic_control/ethnicity_imputation/annual/placebo_highlight_all_imp_date.pdf"))
ggsave(here::here("plots/for_presentation/placebo_highlight_all_imp_date.pdf"), scale = 0.7)


placebo_highlight_all(sc_placebo_pred_full_imp_date_mean, "log_n_pred_full_imp_date")
ggsave(here::here("plots/synthetic_control/ethnicity_imputation/annual/placebo_highlight_all_imp_date_robustnes.pdf"))
ggsave(here::here("plots/for_presentation/placebo_highlight_all_imp_date_robustnes.pdf"), scale = 0.7)


placebo_highlight_mspe(sc_placebo_pred_full_imp_date, "log_n_pred_full_imp_date", exclusion_ratio = 20)
ggsave(here::here("plots/synthetic_control/ethnicity_imputation/annual/placebo_highlight_mspe_20lower_imp_date.pdf"))
ggsave(here::here("plots/for_presentation/placebo_highlight_mspe_20lower_imp_date.pdf"), scale = 0.7)


placebo_highlight_mspe(sc_placebo_pred_full_imp_date_mean, "log_n_pred_full_imp_date", exclusion_ratio = 20)
ggsave(here::here("plots/synthetic_control/ethnicity_imputation/annual/placebo_highlight_mspe_20lower_imp_date_robustnss.pdf"))


placebo_mspe_barplot(sc_placebo_pred_full_imp_date, "log_n_pred_full_imp_date", range_post = c("1933", "1960"))
ggsave(here::here("plots/synthetic_control/ethnicity_imputation/annual/mspe_ratios_imp_date.pdf"))
ggsave(here::here("plots/for_presentation/mspe_ratios_imp_date.pdf"), scale = 0.7)


placebo_mspe_barplot(sc_placebo_pred_full_imp_date, "log_n_pred_full_imp_date", range_post = c("1933", "1939"))
ggsave(here::here("plots/synthetic_control/ethnicity_imputation/annual/mspe_ratios_imp_date_until_1939.pdf"))
ggsave(here::here("plots/for_presentation/mspe_ratios_imp_date_until_1939.pdf"), scale = 0.7)



placebo_mspe_barplot(sc_placebo_pred_full_imp_date_mean, "log_n_pred_full_imp_date", range_post = c("1933", "1960"))
ggsave(here::here("plots/synthetic_control/ethnicity_imputation/annual/mspe_ratios_imp_date_robustness.pdf"))

placebo_mspe_barplot(sc_placebo_pred_full_imp_date_mean, "log_n_pred_full_imp_date", range_post = c("1933", "1939"))
ggsave(here::here("plots/synthetic_control/ethnicity_imputation/annual/mspe_ratios_imp_date_until_1939_robustness.pdf"))

```


```{r comparision plots}
timetk::tk_tbl(sc_results$combined[[dep_var]]) %>% 
  mutate(date = as_date(str_c(index,"-01-01"))) %>% 
  gather("type", dep_var, treated, synth, -date) %>% 
  mutate(type = fct_recode(type, "Synthetic" = "synth", "Actual" = "treated") %>% 
          fct_relevel("Actual")) %>% 
  ggplot(aes(x = date, y = dep_var, col = type)) + theme_minimal() +
  theme(axis.line = element_line(size = 1), 
        #panel.grid.major.x = element_blank(), panel.grid.minor.x = element_blank(), 
        text = element_text(size=12)) + 
  #annotate("rect", fill = "grey", alpha = 0.4, 
  #      xmin = as_date("1921-01-01"), xmax = as_date("1933-01-01"),
  #      ymin = -Inf, ymax = Inf) + 
  geom_line(size = 1) +
  geom_vline(xintercept= ymd(19330101), col = "black", linetype = "dashed", size = 1)+
  scale_x_date(date_minor_breaks = "2 year", date_breaks = "4 year", date_labels = "%Y", 
               limits = c(as_date("1921-01-01"), NA)) +
  labs(x = element_blank(), y = "log(1 + arrests)", col = "German arrests") 

ggsave(here::here("plots/synthetic_control/ethnicity_imputation/annual/comparison_plot.pdf"))



timetk::tk_tbl(sc_results$combined[[dep_var]]) %>% 
  mutate(date = as_date(str_c(index,"-01-01"))) %>% 
  gather("type", dep_var, treated, synth, -date) %>% 
  mutate(type = fct_recode(type, "Synthetic" = "synth", "Actual" = "treated") %>% 
           fct_relevel("Actual"),
         dep_var = exp(dep_var) - 1) %>% 
  ggplot(aes(x = date, y = dep_var, col = type)) + theme_minimal() +
  theme(axis.line = element_line(size = 1), 
        #panel.grid.major.x = element_blank(), panel.grid.minor.x = element_blank(), 
        text = element_text(size=12)) + 
  #annotate("rect", fill = "grey", alpha = 0.4, 
  #      xmin = as_date("1921-01-01"), xmax = as_date("1933-01-01"),
  #      ymin = -Inf, ymax = Inf) + 
  geom_line(size = 1) +
  geom_vline(xintercept= ymd(19330101), col = "black", linetype = "dashed", size = 1)+
  scale_x_date(date_minor_breaks = "2 year", date_breaks = "4 year", date_labels = "%Y", 
               limits = c(as_date("1921-01-01"), NA)) +
  labs(x = element_blank(), y = "Number of arrests", col = "German arrests") +
  scale_y_continuous(trans = "mynaturallog" , breaks = c(0,2,10,100,1000,10000, 100000), labels = comma)

ggsave(here::here("plots/synthetic_control/ethnicity_imputation/annual/comparison_plot_scaled.pdf"))
ggsave(here::here("plots/for_presentation/comparison_plot_scaled.pdf"), scale = 0.7)



opts_current$set(label = "sc_weights")
tibble(ethnicity = names(sc_results$w), Weights = sc_results$w) %>% 
  filter(Weights != 0) %>% 
  arrange(desc(Weights)) %>% 
  kable("latex", booktabs = T, digits = 2, linesep = "", format.args = list(big.mark = " "),
        col.names = c("Ethnic group", "$W$-Weight"), escape = FALSE,
        caption = "Synthetic German minority weights") %>%
 # footnote(general = "Ethnicity and date of arrest were imputed. Full matrix adjustment was applied on ethnic group imputations. All 38 ethnic groups are included.", threeparttable = T) %>% 
    write_file(here::here("tables/sc_weights.tex"))
 

opts_current$set(label = "sc_weights_robustness")
tibble(ethnicity = names(sc_results_mean$w), Weights = sc_results_mean$w) %>% 
  filter(Weights != 0) %>% 
  arrange(desc(Weights)) %>% 
  kable("latex", booktabs = T, digits = 2, linesep = "", format.args = list(big.mark = " "),
        col.names = c("Ethnic group", "$W$-Weight"), escape = FALSE,
        caption = "Synthetic German minority weights") %>%
 # footnote(general = "Ethnicity and date of arrest were imputed. Full matrix adjustment was applied on ethnic group imputations. All 38 ethnic groups are included.", threeparttable = T) %>% 
    write_file(here::here("tables/sc_weights_robustness.tex")) 
  
opts_current$set(label = "sc_weights_robustness")
make_w_weights_table <- function(sc_data, table_title = "Synthetic German minority weights",
                                 file_name = "tables/sc_weights_robustness.tex"){
tibble(ethnicity = names(sc_data$w), Weights = sc_data$w) %>% 
  filter(Weights != 0) %>% 
  arrange(desc(Weights)) %>% 
  kable("latex", booktabs = T, digits = 2, linesep = "", format.args = list(big.mark = " "),
        col.names = c("Ethnic group", "$W$-Weight"), escape = FALSE,
        caption =  table_title) %>%
 # footnote(general = "Ethnicity and date of arrest were imputed. Full matrix adjustment was applied on ethnic group imputations. All 38 ethnic groups are included.", threeparttable = T) %>% 
    write_file(here::here(file_name)) 
}
```

### Robustness checks
Only rehabilitated individuals
```{r}
dep_var <- "log_n_pred_full_imp_date_rehab"
times.dep  <- cbind("log_n_pred_full_imp_date_rehab" = c("1921","1932"))
times.pred <- cbind("log_n_pred_full_imp_date"       = c("1921","1932"),
                    "pop_total"             = c("1921","1932"),
                    "clad_sim"              = c("1921","1932"),
                    "urb_rate"              = c("1921","1932"))
agg.fns <- rep("id", ncol(times.pred))


sc_placebo_rehabs <- mscmt(data_prep_mscmt, treatment.identifier, controls.identifier, times.dep, times.pred, agg.fns, seed=2019, placebo = TRUE, single.v=TRUE)

sc_results_rehabs <- mscmt(data_prep_mscmt, treatment.identifier, controls.identifier,
                                                  times.dep, times.pred, agg.fns, seed=2019,  single.v=TRUE)


placebo_highlight_all(sc_placebo_rehabs, "log_n_pred_full_imp_date_rehab")
ggsave(here::here("plots/final/placebo_highlight_all_rehabs.pdf"))
ggsave(here::here("plots/for_presentation/placebo_highlight_all_rehabs.pdf"), scale = 0.7)


  
opts_current$set(label = "sc_weights_rehabs")
make_w_weights_table(sc_data = sc_results_rehabs, 
                     table_title = "Synthetic German minority weights, Only rehabilitated individuals",
                                 file_name = "tables/sc_weights_rehabs.tex")

placebo_mspe_barplot(sc_placebo_rehabs, 
                     "log_n_pred_full_imp_date_rehab", range_post = c("1933", "1960"))
ggsave(here::here("plots/final/mspe_ratios_rehabs.pdf"))

placebo_mspe_barplot(sc_placebo_rehabs, 
                     "log_n_pred_full_imp_date_rehab", range_post = c("1933", "1939"))
ggsave(here::here("plots/final/mspe_ratios_rehabs_until_1939.pdf"))
```

Only ethnicites without independent state
```{r}
ethnicities_without_ind_state <- ethnicity_controls %>% 
  filter(ind_country != 1) %>% 
  pull(ethnicity)
  
controls.identifier  <- setdiff(ethnicities_without_ind_state,
                                 treatment.identifier)
dep_var <- "log_n_pred_full_imp_date"
times.dep  <- cbind("log_n_pred_full_imp_date" = c("1921","1932"))

sc_placebo_ethnicities_without_ind_state <- mscmt(data_prep_mscmt, treatment.identifier, controls.identifier, times.dep, times.pred, agg.fns, seed=2019, placebo = TRUE, single.v=TRUE)

sc_results_ethnicities_without_ind_state <- mscmt(data_prep_mscmt, treatment.identifier, controls.identifier,
                                                  times.dep, times.pred, agg.fns, seed=2019,  single.v=TRUE)


placebo_highlight_all(sc_placebo_ethnicities_without_ind_state, "log_n_pred_full_imp_date")
ggsave(here::here("plots/final/placebo_highlight_all_ethnicities_without_ind_state.pdf"))
ggsave(here::here("plots/for_presentation/placebo_highlight_all_ethnicities_without_ind_state.pdf"), scale = 0.7)


  
opts_current$set(label = "sc_weights_without_ind_state")
make_w_weights_table(sc_data = sc_results_ethnicities_without_ind_state, 
                     table_title = "Synthetic German minority weights, Only ethnicities without ind. state",
                                 file_name = "tables/sc_weights_without_ind_state.tex")

placebo_mspe_barplot(sc_placebo_ethnicities_without_ind_state, 
                     "log_n_pred_full_imp_date", range_post = c("1933", "1960"))
ggsave(here::here("plots/final/mspe_ratios_ethnicities_without_ind_state.pdf"))

placebo_mspe_barplot(sc_placebo_ethnicities_without_ind_state, 
                     "log_n_pred_full_imp_date", range_post = c("1933", "1939"))
ggsave(here::here("plots/final/mspe_ratios_ethnicities_without_ind_state_until_1939.pdf"))
```



```{r}
dep_var <- "log_n_pred_full"
times.dep  <- cbind("log_n_pred_full" = c("1921","1932"))

sc_placebo_pred_full <- mscmt(data_prep_mscmt, treatment.identifier, controls.identifier, times.dep,
                                       times.pred, agg.fns, seed=2019, placebo = TRUE, single.v=TRUE)

placebo_highlight_all(sc_placebo_pred_full, "log_n_pred_full")
ggsave(here::here("plots/for_presentation/placebo_highlight_all_no_date_imputation.pdf"), scale = 0.7)

placebo_mspe_barplot(sc_placebo_pred_full, 
                     "log_n_pred_full", range_post = c("1933", "1960"))

```
