==> code/10_heterogeneity_by_border.Rmd <==
---
title: "Heterogeneity by border frontiers"
author: "Martin Kosík"
date: "May 2, 2019"
output: github_document
editor_options: 
  chunk_output_type: console
---

```{r setup, echo=T, results='hide', message=F, warning=F}
knitr::opts_chunk$set(echo = TRUE, fig.show = 'hide')
library(tidyverse)
library(broom)
library(stargazer)
library(here)
library(readxl)
library(knitr)
library(kableExtra)
library(estimatr)
library(gghighlight)
library(lubridate)
library(clubSandwich)
library(MSCMT)
library(timetk)
library(scales)
source(here::here("code/functions.R"))
```

## Difference-in-differences
```{r import files}
ethnicity_controls <- read_excel(here::here("data/ethnicity_info.xlsx")) %>%
  mutate(ethnicity_id = as.numeric(1:n()),
         urb_rate_pct = urb_rate * 100)


min_by_year <-  read_csv(here::here("data/min_by_year_borders.csv")) %>% 
  mutate(log_n_pred_full_non_border = log(1 + label_non_border + pred_adj_full_scaled_non_border),
         log_n_pred_full_border = log(1 + label_border + pred_adj_full_scaled_border)) %>% 
  left_join(ethnicity_controls, by = "ethnicity") %>% 
  mutate(german = ifelse(ethnicity == "German", 1, 0),
         post_german = german * ifelse(YEAR >= 1933, 1, 0),
         ethnicity = as.factor(ethnicity),
         YEAR_sq = YEAR^2, 
         pre_treatment = ifelse(YEAR >= 1922 & YEAR < 1933, 1, 0),
         hostility = ifelse(YEAR >= 1933 & YEAR <= 1939, 1, 0), 
         pact = ifelse(YEAR > 1939 & YEAR < 1941, 1, 0), 
         war = ifelse(YEAR >= 1941 & YEAR < 1945, 1, 0),
         post_war = ifelse(YEAR >= 1945, 1, 0)) 
```

```{r geopolitical controls definition}
min_by_year <- min_by_year %>% 
  mutate(geopol_finnish_war = as.numeric(YEAR >= 1939 & YEAR < 1945 & (ethnicity == "Finnish")),
         geopol_finnish_postwar = as.numeric(YEAR >=  1945 & (ethnicity == "Finnish")),
         geopol_baltic_anex = as.numeric(YEAR == 1940 & (ethnicity %in% c("Estonian", "Latvian", "Lithuanian"))),
         geopol_baltic_nazi = as.numeric(YEAR >= 1941 & YEAR <= 1943 &
                                           (ethnicity %in% c("Estonian", "Latvian", "Lithuanian"))),
         geopol_baltic_postwar = as.numeric(YEAR >= 1944 & 
                                           (ethnicity %in% c("Estonian", "Latvian", "Lithuanian"))),
         geopol_polish_war = as.numeric(YEAR == 1939 & (ethnicity == "Polish")),
         geopol_polish_soviet = as.numeric(YEAR == 1940 & (ethnicity  == "Polish")),
         geopol_polish_nazi = as.numeric(YEAR >= 1941 & YEAR < 1945 & (ethnicity == "Polish")),
         geopol_polish_postwar = as.numeric(YEAR >=  1945 & (ethnicity == "Polish")),
         geopol_japnese_war = as.numeric(YEAR >= 1938 & YEAR <= 1939 &  (ethnicity == "Japanese")),
         geopol_japnese_neutrality = as.numeric(YEAR >= 1938 & YEAR <= 1944 &  (ethnicity == "Japanese")),
         geopol_japnese_ww2 = as.numeric(YEAR == 1945 &  (ethnicity == "Japanese")),
         geopol_japnese_postwar = as.numeric(YEAR > 1945 &  (ethnicity == "Japanese")),
         geopol_hungarian_war = as.numeric(YEAR >= 1941 & YEAR < 1945 & (ethnicity == "Hungarian")),
         geopol_hungarian_postwar = as.numeric(YEAR >=  1945 & (ethnicity == "Hungarian")))

```

```{r year dummies}
years <- 1922:1960

year_dummies <- map_dfc(years, ~  if (dplyr::last(years) == .){
                                                    as.numeric(min_by_year$YEAR >= .)} else {
                                                    as.numeric(min_by_year$YEAR == .)}) %>% 
  rename_all(funs( c(paste0("year_" ,years))))

min_by_year <- bind_cols(min_by_year, year_dummies)

```


```{r formulas defined}
geopol_vars <- str_subset(names(min_by_year), "geopol")

border_no_trends_geopol <- as.formula(paste("log_n_border ~ ", 
                         paste0("german:","year_", years, collapse = " + "),  "+ as.factor(YEAR) +
                         ethnicity+ ",  paste0(geopol_vars, collapse = " + ")))


non_border_no_trends_geopol <- as.formula(paste("log_n_non_border ~ ", 
                         paste0("german:","year_", years, collapse = " + "),  "+ as.factor(YEAR) +
                         ethnicity+ ",  paste0(geopol_vars, collapse = " + ")))
```

```{r}
plot_effects_robust_se(formula = border_no_trends_geopol, data = min_by_year) 
ggsave(here::here("plots/final/point_range_robust_cr2_border_provinces.pdf"))

plot_effects_robust_se(formula = border_no_trends_geopol, x_axis_breaks = seq(1922, 1960, 3), 
                       x_labels_angle = 0, x_labels_hjust = 0.5)
ggsave(here::here("plots/for_presentation/point_range_robust_cr2_border_provinces.pdf"), scale = 0.7)



plot_effects_robust_se(formula = non_border_no_trends_geopol, data = min_by_year) 
ggsave(here::here("plots/final/point_range_robust_cr2_non_border_provinces.pdf"))

plot_effects_robust_se(formula = non_border_no_trends_geopol, x_axis_breaks = seq(1922, 1960, 3), 
                       x_labels_angle = 0, x_labels_hjust = 0.5)
ggsave(here::here("plots/for_presentation/point_range_robust_cr2_non_border_provinces.pdf"), scale = 0.7)
```

```{r}
fit_robust_model <- function(formula = fmla_window_pre_treat_pred_full_imp_date, 
                             data = min_by_year, level = 0.95, vcov = "CR2", estimatr = 1){
  if (estimatr != 1){
  model <- lm(formula,  data = data)
  model %>% 
  conf_int(vcov = vcov,  level = level,
          cluster = data$ethnicity,  test = "Satterthwaite") %>% 
  as_tibble(rownames = "term") %>% 
  filter(str_detect(term, "german:")) %>% 
 # mutate(nice_labs = str_replace(term,"german:","") %>% 
 #          fct_relevel("pre_treatment", "hostility", "pact", "war", "post_war")) %>% 
  rename(conf.low = CI_L, conf.high = CI_U, estimate = beta)
  }
  else {
    model_robust <- lm_robust(formula = formula, data = data, clusters = ethnicity, se_type = vcov,
                                ci = TRUE, alpha = 1 - level)
    model_robust %>% 
      estimatr::tidy(model_robust) %>% 
      filter(str_detect(term, "german:")) %>% 
      dplyr::select(term, estimate, SE = std.error,  conf.low, conf.high) %>% 
      as_tibble()
  }
  }
```

```{r}
borders_formulas <- c(border_no_trends_geopol, non_border_no_trends_geopol)

## Position dodge
map(borders_formulas, ~ fit_robust_model(formula = ., vcov = "CR2", estimatr = 0)) %>% 
  bind_rows(.id = "id") %>% 
  mutate(nice_labs = str_replace(term,"german:year_","")
         #,nice_labs = ifelse(as.numeric(nice_labs) == max(as.numeric(nice_labs)), str_c(nice_labs, "+"), nice_labs)
         ) %>% 
  mutate(trends = recode_factor(id, "1" = "Border Region", "2" = "Non-border Region")) %>% 
  ggplot( aes(x = nice_labs, y = estimate, ymin = conf.low, ymax = conf.high, group = 1, col = trends))+ 
  geom_pointrange(position = position_dodge2(width = 0.45))+  geom_hline(yintercept= 0) + theme_minimal() + 
  theme(axis.line = element_line(size = 1), 
        panel.grid.major.x = element_blank(), panel.grid.minor.x = element_blank(), 
        text = element_text(size=14),
        legend.position = "bottom", 
        axis.text.x = element_text(angle = 60, hjust = 1)) +
  labs(x = "Year", 
        #caption = "error bars show 95% confidence intervals \n 
         #                      SE are based on the cluster-robust estimator by Pustejovsky and Tipton (2018)", 
       y = "Coefficient", col = "")

ggsave(here::here("plots/final/point_range_robust_cr2_border_compar.pdf"))

```


## Synthetic control method

```{r}
ethnicity_controls <- read_excel(here::here("data/ethnicity_info.xlsx")) %>%
  mutate(ethnicity_id = as.numeric(1:n()),
         urb_rate_pct = urb_rate * 100)

min_by_year <-  read_csv(here::here("data/min_by_year_borders.csv")) %>% 
  mutate(log_n_pred_full_non_border = log(1 + label_non_border + pred_adj_full_scaled_non_border),
         log_n_pred_full_border = log(1 + label_border + pred_adj_full_scaled_border)) %>% 
  left_join(ethnicity_controls, by = "ethnicity") %>% 
  mutate(german = ifelse(ethnicity == "German", 1, 0),
         post_german = german * ifelse(YEAR >= 1933, 1, 0)) 


```


```{r data preparation}
data_prep_mscmt <- listFromLong(as.data.frame(min_by_year), unit.variable = "ethnicity_id", 
                                time.variable="YEAR", unit.names.variable="ethnicity")

```

```{r definiton of variables}
dep_var <- "log_n_border"

treatment.identifier <- "German"
controls.identifier  <- setdiff(colnames(data_prep_mscmt[[1]]),
                                 treatment.identifier)
times.dep  <- cbind(dep_var                 = c("1921","1932"))
times.dep  <- cbind("log_n_border"                 = c("1921","1932"))

times.pred <- cbind("log_n_border"       = c("1921","1932"),
                    "pop_total"             = c("1921","1932"),
                    "clad_sim"              = c("1921","1932"),
                    "urb_rate"              = c("1921","1932"))

#agg.fns <- rep("mean", ncol(times.pred))                       
agg.fns <- rep("id", ncol(times.pred))
```

```{r border placebo plot}
sc_placebo_border <- mscmt(data_prep_mscmt, treatment.identifier, controls.identifier, 
                              times.dep, times.pred, agg.fns,
                              seed=2019, placebo = TRUE, single.v=TRUE)

placebo_highlight_all(sc_placebo_border, "log_n_border")
ggsave(here::here("plots/final/placebo_highlight_all_border_provinces.pdf"))
ggsave(here::here("plots/for_presentation/placebo_highlight_all_border_provinces.pdf"), scale = 0.7)



placebo_highlight_mspe(sc_placebo_border, "log_n_border", exclusion_ratio = 20)

placebo_mspe_barplot(sc_placebo_border, "log_n_border", range_post = c("1933", "1960"))
ggsave(here::here("plots/final/mspe_ratios_border_provinces.pdf"))

placebo_mspe_barplot(sc_placebo_border, "log_n_border", range_post = c("1933", "1939"))
ggsave(here::here("plots/final/mspe_ratios_border_provinces_until_1939.pdf"))


```

```{r non-border placebo plot}
dep_var <- "log_n_non_border"
times.dep  <- cbind("log_n_non_border"      = c("1921","1932"))

times.pred <- cbind("log_n_non_border"      = c("1921","1932"),
                    "pop_total"             = c("1921","1932"),
                    "clad_sim"              = c("1921","1932"),
                    "urb_rate"              = c("1921","1932"))

sc_placebo_non_border <- mscmt(data_prep_mscmt, treatment.identifier, controls.identifier, 
                              times.dep, times.pred, agg.fns,
                              seed=2019, placebo = TRUE, single.v=TRUE)

placebo_highlight_all(sc_placebo_non_border, "log_n_non_border")
ggsave(here::here("plots/final/placebo_highlight_all_non_border_provinces.pdf"))
ggsave(here::here("plots/for_presentation/placebo_highlight_all_non_border_provinces.pdf"), scale = 0.7)



placebo_highlight_mspe(sc_placebo_non_border, "log_n_non_border", exclusion_ratio = 20)


placebo_mspe_barplot(sc_placebo_non_border, "log_n_non_border", range_post = c("1933", "1960"))
ggsave(here::here("plots/final/mspe_ratios_non_border_provinces.pdf"))

placebo_mspe_barplot(sc_placebo_non_border, "log_n_non_border", range_post = c("1933", "1939"))
ggsave(here::here("plots/final/mspe_ratios_non_border_provinces_until_1939.pdf"))


```



# Only western border


## Difference-in-differences
```{r import files}
ethnicity_controls <- read_excel(here::here("data/ethnicity_info.xlsx")) %>%
  mutate(ethnicity_id = as.numeric(1:n()),
         urb_rate_pct = urb_rate * 100)


min_by_year <-  read_csv(here::here("data/min_by_year_borders_only_western.csv")) %>% 
  mutate(log_n_pred_full_non_border = log(1 + label_non_border + pred_adj_full_scaled_non_border),
         log_n_pred_full_border = log(1 + label_border + pred_adj_full_scaled_border)) %>% 
  left_join(ethnicity_controls, by = "ethnicity") %>% 
  mutate(german = ifelse(ethnicity == "German", 1, 0),
         post_german = german * ifelse(YEAR >= 1933, 1, 0),
         ethnicity = as.factor(ethnicity),
         YEAR_sq = YEAR^2, 
         pre_treatment = ifelse(YEAR >= 1922 & YEAR < 1933, 1, 0),
         hostility = ifelse(YEAR >= 1933 & YEAR <= 1939, 1, 0), 
         pact = ifelse(YEAR > 1939 & YEAR < 1941, 1, 0), 
         war = ifelse(YEAR >= 1941 & YEAR < 1945, 1, 0),
         post_war = ifelse(YEAR >= 1945, 1, 0)) 
```

```{r geopolitical controls definition}
min_by_year <- min_by_year %>% 
  mutate(geopol_finnish_war = as.numeric(YEAR >= 1939 & YEAR < 1945 & (ethnicity == "Finnish")),
         geopol_finnish_postwar = as.numeric(YEAR >=  1945 & (ethnicity == "Finnish")),
         geopol_baltic_anex = as.numeric(YEAR == 1940 & (ethnicity %in% c("Estonian", "Latvian", "Lithuanian"))),
         geopol_baltic_nazi = as.numeric(YEAR >= 1941 & YEAR <= 1943 &
                                           (ethnicity %in% c("Estonian", "Latvian", "Lithuanian"))),
         geopol_baltic_postwar = as.numeric(YEAR >= 1944 & 
                                           (ethnicity %in% c("Estonian", "Latvian", "Lithuanian"))),
         geopol_polish_war = as.numeric(YEAR == 1939 & (ethnicity == "Polish")),
         geopol_polish_soviet = as.numeric(YEAR == 1940 & (ethnicity  == "Polish")),
         geopol_polish_nazi = as.numeric(YEAR >= 1941 & YEAR < 1945 & (ethnicity == "Polish")),
         geopol_polish_postwar = as.numeric(YEAR >=  1945 & (ethnicity == "Polish")),
         geopol_japnese_war = as.numeric(YEAR >= 1938 & YEAR <= 1939 &  (ethnicity == "Japanese")),
         geopol_japnese_neutrality = as.numeric(YEAR >= 1938 & YEAR <= 1944 &  (ethnicity == "Japanese")),
         geopol_japnese_ww2 = as.numeric(YEAR == 1945 &  (ethnicity == "Japanese")),
         geopol_japnese_postwar = as.numeric(YEAR > 1945 &  (ethnicity == "Japanese")),
         geopol_hungarian_war = as.numeric(YEAR >= 1941 & YEAR < 1945 & (ethnicity == "Hungarian")),
         geopol_hungarian_postwar = as.numeric(YEAR >=  1945 & (ethnicity == "Hungarian")))

```

```{r year dummies}
years <- 1922:1960

year_dummies <- map_dfc(years, ~  if (dplyr::last(years) == .){
                                                    as.numeric(min_by_year$YEAR >= .)} else {
                                                    as.numeric(min_by_year$YEAR == .)}) %>% 
  rename_all(funs( c(paste0("year_" ,years))))

min_by_year <- bind_cols(min_by_year, year_dummies)

```


```{r formulas defined}
geopol_vars <- str_subset(names(min_by_year), "geopol")

border_no_trends_geopol <- as.formula(paste("log_n_border ~ ", 
                         paste0("german:","year_", years, collapse = " + "),  "+ as.factor(YEAR) +
                         ethnicity+ ",  paste0(geopol_vars, collapse = " + ")))


non_border_no_trends_geopol <- as.formula(paste("log_n_non_border ~ ", 
                         paste0("german:","year_", years, collapse = " + "),  "+ as.factor(YEAR) +
                         ethnicity+ ",  paste0(geopol_vars, collapse = " + ")))
```

```{r}

plot_effects_robust_se(formula = border_no_trends_geopol, x_axis_breaks = seq(1922, 1960, 3), 
                       x_labels_angle = 0, x_labels_hjust = 0.5)
ggsave(here::here("plots/for_presentation/point_range_robust_cr2_border_provinces_western.pdf"), scale = 0.7)



plot_effects_robust_se(formula = non_border_no_trends_geopol, x_axis_breaks = seq(1922, 1960, 3), 
                       x_labels_angle = 0, x_labels_hjust = 0.5)
ggsave(here::here("plots/for_presentation/point_range_robust_cr2_non_border_provinces_western.pdf"), scale = 0.7)
```


## Synthetic control method

```{r}
ethnicity_controls <- read_excel(here::here("data/ethnicity_info.xlsx")) %>%
  mutate(ethnicity_id = as.numeric(1:n()),
         urb_rate_pct = urb_rate * 100)

min_by_year <-  read_csv(here::here("data/min_by_year_borders_only_western.csv")) %>% 
  mutate(log_n_pred_full_non_border = log(1 + label_non_border + pred_adj_full_scaled_non_border),
         log_n_pred_full_border = log(1 + label_border + pred_adj_full_scaled_border)) %>% 
  left_join(ethnicity_controls, by = "ethnicity") %>% 
  mutate(german = ifelse(ethnicity == "German", 1, 0),
         post_german = german * ifelse(YEAR >= 1933, 1, 0)) 


```


```{r data preparation}
data_prep_mscmt <- listFromLong(as.data.frame(min_by_year), unit.variable = "ethnicity_id", 
                                time.variable="YEAR", unit.names.variable="ethnicity")

```

```{r definiton of variables}
dep_var <- "log_n_border"

treatment.identifier <- "German"
controls.identifier  <- setdiff(colnames(data_prep_mscmt[[1]]),
                                 treatment.identifier)
times.dep  <- cbind(dep_var                 = c("1921","1932"))
times.dep  <- cbind("log_n_border"                 = c("1921","1932"))

times.pred <- cbind("log_n_border"       = c("1921","1932"),
                    "pop_total"             = c("1921","1932"),
                    "clad_sim"              = c("1921","1932"),
                    "urb_rate"              = c("1921","1932"))

#agg.fns <- rep("mean", ncol(times.pred))                       
agg.fns <- rep("id", ncol(times.pred))
```

```{r border placebo plot}
sc_placebo_border <- mscmt(data_prep_mscmt, treatment.identifier, controls.identifier, 
                              times.dep, times.pred, agg.fns,
                              seed=2019, placebo = TRUE, single.v=TRUE)

placebo_highlight_all(sc_placebo_border, "log_n_border")
ggsave(here::here("plots/for_presentation/placebo_highlight_all_border_provinces_western.pdf"), scale = 0.7)


```

```{r non-border placebo plot}
dep_var <- "log_n_non_border"
times.dep  <- cbind("log_n_non_border"      = c("1921","1932"))

times.pred <- cbind("log_n_non_border"      = c("1921","1932"),
                    "pop_total"             = c("1921","1932"),
                    "clad_sim"              = c("1921","1932"),
                    "urb_rate"              = c("1921","1932"))

sc_placebo_non_border <- mscmt(data_prep_mscmt, treatment.identifier, controls.identifier, 
                              times.dep, times.pred, agg.fns,
                              seed=2019, placebo = TRUE, single.v=TRUE)

placebo_highlight_all(sc_placebo_non_border, "log_n_non_border")
ggsave(here::here("plots/for_presentation/placebo_highlight_all_non_border_provinces_western.pdf"), scale = 0.7)



```
